util/fi.h: Move network related functions to new header file
Isolate general network defines into new header, ofi_net.h.  This
makes it easier to find generic network routines.
